Finish file downloading and stream pausing. Add file download tests

+++ src/tests/utils.h
@@ -10,6 +10,6 @@
 using TestWsServer = SimpleWeb::SocketServer<SimpleWeb::WS>;
 
 auto randomInt(uint64_t start, uint64_t end) -> uint64_t;
-auto generateRandomData(uint32_t count) -> std::shared_ptr<std::vector<uint8_t>>;
+auto generateRandomData(size_t count) -> std::shared_ptr<std::vector<uint8_t>>;
 
 #endif //ADACS_JOB_CLIENT_UTILS_H
